requirements.txt
gradio
httpx
pytorch_tabular
aiofiles=23.2.1=pypi_0
aiohttp=3.9.1=pypi_0
aiosignal=1.3.1=pypi_0
altair=5.2.0=pypi_0
annotated-types=0.6.0=pypi_0
anyio=3.7.1=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.3=pypi_0
azure-core=1.29.5=pypi_0
azure-identity=1.15.0=pypi_0
azure-storage-blob=12.19.0=pypi_0
backcall=0.2.0=pyhd3eb1b0_0
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.08.22=h06a4308_0
certifi=2023.11.17=pypi_0
click=8.1.7=pypi_0
codecarbon=2.3.2=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.2=py310h06a4308_0
contourpy=1.2.0=pypi_0
cryptography=41.0.7=pypi_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py310h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
executing=0.8.3=pyhd3eb1b0_0
fastapi=0.104.1=pypi_0
ffmpy=0.3.1=pypi_0
filelock=3.13.1=pypi_0
fonttools=4.45.1=pypi_0
frozenlist=1.4.0=pypi_0
fsspec=2023.10.0=pypi_0
fuzzywuzzy=0.18.0=pypi_0
grad-cam=1.4.8=pypi_0
gradio=4.7.1=pypi_0
gradio-client=0.7.0=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.2=pypi_0
httpx=0.25.2=pypi_0
huggingface-hub=0.19.4=pypi_0
idna=3.6=pypi_0
importlib-metadata=6.0.0=py310h06a4308_0
importlib-resources=6.1.1=pypi_0
importlib_metadata=6.0.0=hd3eb1b0_0
ipykernel=6.25.0=py310h2f386ee_0
ipython=8.12.2=py310h06a4308_0
isodate=0.6.1=pypi_0
jedi=0.18.1=py310h06a4308_1
joblib=1.3.2=pypi_0
jupyter_client=8.6.0=py310h06a4308_0
jupyter_core=5.5.0=py310h06a4308_0
kiwisolver=1.4.5=pypi_0
ld_impl_linux-64=2.38=h1181459_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.3=pypi_0
matplotlib=3.8.2=pypi_0
matplotlib-inline=0.1.6=py310h06a4308_0
mdurl=0.1.2=pypi_0
mpmath=1.3.0=pypi_0
msal=1.25.0=pypi_0
msal-extensions=1.0.0=pypi_0
multidict=6.0.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py310h06a4308_0
networkx=3.2.1=pypi_0
nvidia-cublas-cu12=12.1.3.1=pypi_0
nvidia-cuda-cupti-cu12=12.1.105=pypi_0
nvidia-cuda-nvrtc-cu12=12.1.105=pypi_0
nvidia-cuda-runtime-cu12=12.1.105=pypi_0
nvidia-cudnn-cu12=8.9.2.26=pypi_0
nvidia-cufft-cu12=11.0.2.54=pypi_0
nvidia-curand-cu12=10.3.2.106=pypi_0
nvidia-cusolver-cu12=11.4.5.107=pypi_0
nvidia-cusparse-cu12=12.1.0.106=pypi_0
nvidia-nccl-cu12=2.18.1=pypi_0
nvidia-nvjitlink-cu12=12.3.101=pypi_0
nvidia-nvtx-cu12=12.1.105=pypi_0
openssl=3.0.12=h7f8727e_0
orjson=3.9.10=pypi_0
packaging=23.1=py310h06a4308_0
pandas=2.1.3=pypi_0
parso=0.8.3=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.1.0=pypi_0
pip=23.3.1=py310h06a4308_0
platformdirs=3.10.0=py310h06a4308_0
portalocker=2.8.2=pypi_0
prompt-toolkit=3.0.36=py310h06a4308_0
psutil=5.9.0=py310h5eee18b_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py-cpuinfo=9.0.0=pypi_0
pydantic=2.5.2=pypi_0
pydantic-core=2.14.5=pypi_0
pydub=0.25.1=pypi_0
pygments=2.15.1=py310h06a4308_1
pyjwt=2.8.0=pypi_0
pynvml=11.5.0=pypi_0
pyparsing=3.1.1=pypi_0
python=3.10.13=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.0=pypi_0
python-multipart=0.0.6=pypi_0
pytz=2023.3.post1=pypi_0
pyyaml=6.0.1=pypi_0
pyzmq=25.1.0=py310h6a678d5_0
readline=8.2=h5eee18b_0
rich=13.7.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.11.4=pypi_0
semantic-version=2.10.0=pypi_0
setuptools=68.0.0=py310h06a4308_0
shellingham=1.5.4=pypi_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
starlette=0.27.0=pypi_0
sympy=1.12=pypi_0
threadpoolctl=3.2.0=pypi_0
tk=8.6.12=h1ccaba5_0
tomlkit=0.12.0=pypi_0
toolz=0.12.0=pypi_0
torch=2.1.1=pypi_0
torchaudio=2.1.1=pypi_0
torchvision=0.16.1=pypi_0
tornado=6.3.3=py310h5eee18b_0
tqdm=4.66.1=pypi_0
traitlets=5.7.1=py310h06a4308_0
triton=2.1.0=pypi_0
ttach=0.0.3=pypi_0
typer=0.9.0=pypi_0
tzdata=2023.3=pypi_0
urllib3=2.1.0=pypi_0
uvicorn=0.24.0.post1=pypi_0
wcwidth=0.2.5=pyhd3eb1b0_0
websockets=11.0.3=pypi_0
wheel=0.41.2=py310h06a4308_0
xz=5.4.2=h5eee18b_0
yarl=1.9.3=pypi_0
zeromq=4.3.4=h2531618_0
zipp=3.11.0=py310h06a4308_0
zlib=1.2.13=h5eee18b_0
